# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/bark/{ticket}/push": {
            "post": {
                "description": "接收Bark消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Adapter"
                ],
                "summary": "接收Bark消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket",
                        "name": "ticket",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Subtitle",
                        "name": "subtitle",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Body",
                        "name": "body",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/api/gotify/message": {
            "post": {
                "description": "接收Gotify消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Adapter"
                ],
                "summary": "接收Gotify消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Message",
                        "name": "message",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/api/message/list": {
            "get": {
                "description": "获取所有消息的列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "获取所有消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Message"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/api/ntfy/{ticket}": {
            "post": {
                "description": "接收Ntfy消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Adapter"
                ],
                "summary": "接收Ntfy消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket",
                        "name": "ticket",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Message",
                        "name": "message",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/onebot/{ticket}/send_msg": {
            "post": {
                "description": "接收OneBot消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Adapter"
                ],
                "summary": "接收OneBot消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket",
                        "name": "ticket",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Message",
                        "name": "message",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/pushdeer/message/push": {
            "post": {
                "description": "接收PushDeer消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Adapter"
                ],
                "summary": "接收PushDeer消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text",
                        "name": "text",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Desp",
                        "name": "desp",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/webhook/{ticket}": {
            "post": {
                "description": "接收Webhook消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Adapter"
                ],
                "summary": "接收Webhook消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket",
                        "name": "ticket",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Message",
                        "name": "message",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Bridge": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "source_channel_type": {
                    "$ref": "#/definitions/types.ChannelType"
                },
                "target_channel": {
                    "$ref": "#/definitions/models.Channel"
                },
                "target_channel_id": {
                    "type": "integer"
                },
                "ticket": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Channel": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "渠道配置（统一结构）",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "description": "UUID",
                    "type": "integer"
                },
                "name": {
                    "description": "渠道名称（如：Email-SMTP、OneBot-v11）",
                    "type": "string"
                },
                "type": {
                    "description": "渠道类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.ChannelType"
                        }
                    ]
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Message": {
            "type": "object",
            "properties": {
                "bridge": {
                    "$ref": "#/definitions/models.Bridge"
                },
                "bridge_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/types.MessageStatus"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "types.ChannelType": {
            "type": "string",
            "enum": [
                "Bark",
                "DingTalk",
                "FeiShu",
                "OneBot",
                "Gotify",
                "ServerChan",
                "PushDeer",
                "Synology",
                "IYUU",
                "PushPlus",
                "QMSG",
                "WeCom",
                "Telegram",
                "Email",
                "Webhook",
                "Ntfy",
                "WxPusher"
            ],
            "x-enum-comments": {
                "TypeBark": "Bark",
                "TypeDingTalk": "钉钉机器人",
                "TypeEmail": "邮箱",
                "TypeFeiShu": "飞书机器人",
                "TypeGotify": "Gotify",
                "TypeIYUU": "IYUU",
                "TypeNtfy": "Ntfy",
                "TypeOneBot": "OneBot",
                "TypePushDeer": "PushDeer",
                "TypePushPlus": "PushPlus",
                "TypeQMSG": "Qmsg酱",
                "TypeServerChan": "Server酱",
                "TypeSynology": "群晖chat",
                "TypeTelegram": "Telegram",
                "TypeWeCom": "企业微信",
                "TypeWebhook": "Webhook",
                "TypeWxPusher": "WxPusher"
            },
            "x-enum-varnames": [
                "TypeBark",
                "TypeDingTalk",
                "TypeFeiShu",
                "TypeOneBot",
                "TypeGotify",
                "TypeServerChan",
                "TypePushDeer",
                "TypeSynology",
                "TypeIYUU",
                "TypePushPlus",
                "TypeQMSG",
                "TypeWeCom",
                "TypeTelegram",
                "TypeEmail",
                "TypeWebhook",
                "TypeNtfy",
                "TypeWxPusher"
            ]
        },
        "types.MessageStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-comments": {
                "StatusFailed": "失败",
                "StatusPending": "待发送",
                "StatusSending": "发送中",
                "StatusSuccess": "已发送"
            },
            "x-enum-varnames": [
                "StatusPending",
                "StatusSending",
                "StatusSuccess",
                "StatusFailed"
            ]
        }
    }
}